+++ src/main/java/com/jsunsoft/http/RetryableWebTarget.java
@@ -27,8 +27,8 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.time.Duration;
 import java.util.Collection;
-import java.util.concurrent.TimeUnit;
 
 @Beta
 class RetryableWebTarget extends BasicWebTarget {
@@ -50,17 +50,28 @@ class RetryableWebTarget extends BasicWebTarget {
     public Response request(HttpMethod method, HttpContext context) {
         Response response = super.request(method, context);
 
-        int retryCount = retryContext.getRetryCount();
+        final int maxRetries = retryContext.getRetryCount();
+        int remaining = maxRetries;
+        int attemptNumber = 1;
 
         try {
-            while (retryCount > 0 && retryContext.mustBeRetried(response)) {
-                LOGGER.debug("Request to URI: [{}] has been retried. Response code: [{}]", response.getURI(), response.getCode());
+            while (remaining > 0) {
+                RetryAttempt attempt = new BasicRetryAttempt(response, method, response.getURI(), attemptNumber, null);
+                if (!retryContext.mustBeRetried(attempt)) {
+                    break;
+                }
 
-                TimeUnit.SECONDS.sleep(retryContext.getRetryDelay(response));
+                LOGGER.debug("Request to URI: [{}] will be retried (attempt {} of {}). Response code: [{}]",
+                        response.getURI(), attemptNumber + 1, maxRetries + 1, response.getCode());
+
+                Duration delay = retryContext.getRetryDelay(attempt);
+                if (delay != null && !delay.isZero() && !delay.isNegative()) {
+                    Thread.sleep(delay.toMillis());
+                }
 
                 closeResponse(response);
 
-                WebTarget retryTarget = retryContext.beforeRetry(this);
+                WebTarget retryTarget = retryContext.beforeRetry(attempt, this);
                 if (retryTarget instanceof RetryableWebTarget) {
                     // Avoid recursion (and retryCount reset) when beforeRetry returns the same retryable instance.
                     // Execute a single request attempt using a non-retryable target copy.
@@ -68,7 +79,8 @@ class RetryableWebTarget extends BasicWebTarget {
                 } else {
                     response = retryTarget.request(method, context);
                 }
-                retryCount--;
+                attemptNumber++;
+                remaining--;
             }
         } catch (InterruptedException e) {
             closeResponse(response);